View which holds all steps should aggregate datapackage from child views (#27)
