+++ os-import/src/scripts/components/ui/create-dp/index.js
@@ -4,6 +4,7 @@ var backbone = require('backbone');
 var DataFilesFormView = require('./data-files-form');
 var HeaderView = require('./header');
 var MapperView = require('./mapper');
+var slug = require('slug');
 
 module.exports = backbone.BaseView.extend({
   activate: function(state) {
@@ -17,6 +18,26 @@ module.exports = backbone.BaseView.extend({
   deactivate: function() {
     _.invoke(_.values(this.layout), 'activate', false);
     return this;
+  },
+
+  getDatapackage: function() {
+    var value = this.layout.form.getValue();
+
+    return JSON.parse(window.TEMPLATES['datapackage.hbs']({
+      name: slug(value.name).toLowerCase(),
+      title: value.name,
+
+      resources: _.map(value.files, function(file) {
+        var filePath = file.isURL ? _.last(file.name.split('/')) : file.name;
+
+        return {
+          bytes   : file.size,
+          filename: _.initial(filePath.split('.')).join('.'),
+          schema  : JSON.stringify(file.schema),
+          path    : filePath
+        };
+      })
+    }));
   },
 
   render: function() {
